+++ libs/client/src/person.rs
@@ -1,9 +1,9 @@
 use std::collections::BTreeMap;
 
 use kanidm_proto::constants::*;
-use kanidm_proto::internal::{IdentifyUserRequest, IdentifyUserResponse};
+use kanidm_proto::internal::{CredentialStatus, IdentifyUserRequest, IdentifyUserResponse};
 use kanidm_proto::v1::{
-    AccountUnixExtend, CredentialStatus, Entry, SingleStringRequest, UatStatus,
+    AccountUnixExtend, Entry, SingleStringRequest, UatStatus,
 };
 use uuid::Uuid;
