+++ proto/src/internal/error.rs
@@ -1,0 +1,147 @@
+use uuid::Uuid;
+use serde::{Deserialize, Serialize};
+use utoipa::ToSchema;
+
+use super::credupdate::PasswordFeedback;
+
+/* ===== errors ===== */
+#[derive(Serialize, Deserialize, Debug, PartialEq, Eq, ToSchema)]
+#[serde(rename_all = "lowercase")]
+pub enum SchemaError {
+    NotImplemented,
+    NoClassFound,
+    InvalidClass(Vec<String>),
+    MissingMustAttribute(Vec<String>),
+    InvalidAttribute(String),
+    InvalidAttributeSyntax(String),
+    AttributeNotValidForClass(String),
+    SupplementsNotSatisfied(Vec<String>),
+    ExcludesNotSatisfied(Vec<String>),
+    EmptyFilter,
+    Corrupted,
+    PhantomAttribute(String),
+}
+
+#[derive(Serialize, Deserialize, Debug, PartialEq, Eq, ToSchema)]
+#[serde(rename_all = "lowercase")]
+pub enum PluginError {
+    AttrUnique(String),
+    Base(String),
+    ReferentialIntegrity(String),
+    CredImport(String),
+    Oauth2Secrets,
+}
+
+#[derive(Serialize, Deserialize, Debug, PartialEq, Eq, ToSchema)]
+#[serde(rename_all = "lowercase")]
+pub enum ConsistencyError {
+    Unknown,
+    // Class, Attribute
+    SchemaClassMissingAttribute(String, String),
+    SchemaClassPhantomAttribute(String, String),
+    SchemaUuidNotUnique(Uuid),
+    QueryServerSearchFailure,
+    EntryUuidCorrupt(u64),
+    UuidIndexCorrupt(String),
+    UuidNotUnique(String),
+    RefintNotUpheld(u64),
+    MemberOfInvalid(u64),
+    InvalidAttributeType(String),
+    DuplicateUniqueAttribute,
+    InvalidSpn(u64),
+    SqliteIntegrityFailure,
+    BackendAllIdsSync,
+    BackendIndexSync,
+    ChangelogDesynchronised(u64),
+    ChangeStateDesynchronised(u64),
+    RuvInconsistent(String),
+    DeniedName(Uuid),
+}
+
+#[derive(Serialize, Deserialize, Debug, ToSchema)]
+#[serde(rename_all = "lowercase")]
+pub enum OperationError {
+    SessionExpired,
+    EmptyRequest,
+    Backend,
+    NoMatchingEntries,
+    NoMatchingAttributes,
+    CorruptedEntry(u64),
+    CorruptedIndex(String),
+    // TODO: this should just be a vec of the ConsistencyErrors, surely?
+    ConsistencyError(Vec<Result<(), ConsistencyError>>),
+    SchemaViolation(SchemaError),
+    Plugin(PluginError),
+    FilterGeneration,
+    FilterUuidResolution,
+    InvalidAttributeName(String),
+    InvalidAttribute(String),
+    InvalidDbState,
+    InvalidCacheState,
+    InvalidValueState,
+    InvalidEntryId,
+    InvalidRequestState,
+    InvalidSyncState,
+    InvalidState,
+    InvalidEntryState,
+    InvalidUuid,
+    InvalidReplChangeId,
+    InvalidAcpState(String),
+    InvalidSchemaState(String),
+    InvalidAccountState(String),
+    MissingEntries,
+    ModifyAssertionFailed,
+    BackendEngine,
+    SqliteError, //(RusqliteError)
+    FsError,
+    SerdeJsonError,
+    SerdeCborError,
+    AccessDenied,
+    NotAuthenticated,
+    NotAuthorised,
+    InvalidAuthState(String),
+    InvalidSessionState,
+    SystemProtectedObject,
+    SystemProtectedAttribute,
+    PasswordQuality(Vec<PasswordFeedback>),
+    CryptographyError,
+    ResourceLimit,
+    QueueDisconnected,
+    Webauthn,
+    #[serde(with = "time::serde::timestamp")]
+    Wait(time::OffsetDateTime),
+    ReplReplayFailure,
+    ReplEntryNotChanged,
+    ReplInvalidRUVState,
+    ReplDomainLevelUnsatisfiable,
+    ReplDomainUuidMismatch,
+    ReplServerUuidSplitDataState,
+    TransactionAlreadyCommitted,
+    /// when you ask for a gid that's lower than a safe minimum
+    GidOverlapsSystemMin(u32),
+    /// When a name is denied by the system config
+    ValueDenyName,
+    // What about something like this for unique errors?
+    // Credential Update Errors
+    CU0001WebauthnAttestationNotTrusted,
+    CU0002WebauthnRegistrationError,
+    // ValueSet errors
+    VS0001IncomingReplSshPublicKey,
+    // Value Errors
+    VL0001ValueSshPublicKeyString,
+    // SCIM
+    SC0001IncomingSshPublicKey,
+    // Migration
+    MG0001InvalidReMigrationLevel,
+    MG0002RaiseDomainLevelExceedsMaximum,
+    MG0003ServerPhaseInvalidForMigration,
+}
+
+impl PartialEq for OperationError {
+    fn eq(&self, other: &Self) -> bool {
+        // We do this to avoid InvalidPassword being checked as it's not
+        // derive PartialEq. Generally we only use the PartialEq for TESTING
+        // anyway.
+        std::mem::discriminant(self) == std::mem::discriminant(other)
+    }
+}
