+++ proto/src/lib.rs
@@ -1,5 +1,10 @@
 #![deny(warnings)]
 #![warn(unused_extern_crates)]
+
+// #![warn(missing_docs)]
+
+#![deny(clippy::todo)]
+#![deny(clippy::unimplemented)]
 #![deny(clippy::unwrap_used)]
 #![deny(clippy::expect_used)]
 #![deny(clippy::panic)]
