+++ proto/src/scim_v1/mod.rs
@@ -1,0 +1,3 @@
+mod synch;
+
+pub use self::synch::*;
