+++ server/web_ui/shared/src/models.rs
@@ -2,7 +2,7 @@
 use gloo::console;
 use gloo::storage::{LocalStorage, SessionStorage as TemporaryStorage, Storage};
 use kanidm_proto::oauth2::AuthorisationRequest;
-use kanidm_proto::v1::{CUSessionToken, CUStatus};
+use kanidm_proto::internal::{CUSessionToken, CUStatus};
 use wasm_bindgen::UnwrapThrowExt;
 
 use crate::constants::URL_USER_HOME;
